combine cell_share and sharing_components
